+++ x/cronos/handler.go
@@ -32,6 +32,9 @@ func NewHandler(k keeper.Keeper) sdk.Handler {
 		case *types.MsgTurnBridge:
 			res, err := msgServer.TurnBridge(sdk.WrapSDKContext(ctx), msg)
 			return sdk.WrapServiceResult(ctx, res, err)
+		case *types.MsgStoreBlockList:
+			res, err := msgServer.StoreBlockList(sdk.WrapSDKContext(ctx), msg)
+			return sdk.WrapServiceResult(ctx, res, err)
 		default:
 			errMsg := fmt.Sprintf("unrecognized %s message type: %T", types.ModuleName, msg)
 			return nil, errors.Wrap(sdkerrors.ErrUnknownRequest, errMsg)